update the reconnection example [skip ci]

+++ examples/reconnection_2d/diffusion_reconnection.sh
@@ -3,7 +3,10 @@
 MACHINE=Perlmutter
 # MACHINE=Frontera
 
-conf=conf.dat
+quota_hour=12.0
+restart_flag=.false.  # whether to restart a previous simulation
+
+conf=conf_reconnection.dat
 mhd_config_filename=mhd_config.dat
 
 focused_transport=.false.
@@ -12,9 +15,11 @@ mpi_size=128
 ntasks_per_node=128
 size_mpi_sub=1   # Size of a MPI sub-communicator
 nptl=500
+nptl_max=1000000
 ts=0
 te=200
-single_time_frame=0
+tmax_mhd=200         # the maximum time frame for the MHD fields
+single_time_frame=0  # whether to use only a single MHD frame
 time_interp=1        # whether to interpolation in-between time frames
 dist_flag=1          # 0 for Maxwellian. 1 for delta function. 2 for power-law
 power_index=6.2      # power-law index for the initial distribution
@@ -22,15 +27,16 @@ split_flag=1         # 0 for without particle split, 1 for with split
 split_ratio=2.0      # (> 1.0 ) momentum increase ratio for particle splitting
 pmin_split=2.0       # (> 1.0) minimum momentum (in p0) to start splitting particles
 local_dist=.true.    # whether diagnose local particle distribution
-dump_escaped=.false. # whether to dump escaped particles
+dump_escaped_dist=.true.  # whether to dump escaped particle distributions
+dump_escaped=.false.  # whether to dump escaped particles
 
-track_particle_flag=0 # whether to track particles
-nptl_selected=100     # number of selected particles to track
-nsteps_interval=1   # steps interval to track particles
-nptl_max=1000000
+track_particle_flag=.false. # whether to track particles
+particle_tags_file=tags_selected_01.h5 # HDF5 file containing particle tags selected
+nsteps_interval=100           # steps interval to track particles
 
 inject_new_ptl=1 # whether to inject new particles at every step
 inject_same_nptl=1 # whether to inject the same number of particles every step
+tmax_to_inject=200  # the maximum time frame to inject particles
 inject_part_box=0 # inject in part of the box
 ptl_xmin=-0.25
 ptl_xmax=0.25
@@ -58,9 +64,6 @@ weak_scattering=1 # whether in weak-scattering regime
 deltab_flag=0   # whether to have spatially dependent turbulence amplitude
 correlation_flag=0 # whether to have spatially dependent turbulence correlation length
 ndim_field=2 # The dimension of the field
-drift_param1=850964.408   # parameter #1 for particle drift in 3D
-drift_param2=13575468.975 # parameter #2 for particle drift in 3D
-charge=-1 # charge in unit charge
 spherical_coord=0 # whether the grid is spherical
 uniform_grid=1 # whether the grid is uniform
 check_drift_2d=0 # whether to check particle drift in 2D
@@ -83,10 +86,12 @@ change_variable () {
 }
 
 if [ "$MACHINE" = "Perlmutter" ]; then
+    module load cpu cray-hdf5-parallel
     export OMP_PROC_BIND=true
     export OMP_PLACES=cores
     export OMP_NUM_THREADS=2
 else
+    module load phdf5
     export IBRUN_TASKS_PER_NODE=$ntasks_per_node
     export OMP_NUM_THREADS=1
 fi
@@ -101,23 +106,32 @@ run_stochastic () {
     cd ..
     diagnostics_directory=$7
     mkdir -p $diagnostics_directory
-    rm $diagnostics_directory/*
+    if [ "$restart_flag" = ".false." ] && [ "$track_particle_flag" = ".false." ]; then
+        rm $diagnostics_directory/*
+    fi
+    mkdir -p $diagnostics_directory/restart
     tau0_scattering=$8
     duu0=$9
     particle_v0=${10}
+    drift_param1=${11}
+    drift_param2=${12}
+    charge=${13}
     commands="./stochastic-mhd.exec \
+        -qh $quota_hour -rf $restart_flag \
         -ft $focused_transport -pv $particle_v0 \
         -sm $size_mpi_sub \
         -dm $4 -mc $mhd_config_filename -np $nptl \
-        -ti $time_interp -ts $ts -te $te -st $single_time_frame \
+        -ti $time_interp -ts $ts -te $te -tm $tmax_to_inject \
+        -st $single_time_frame \
         -df $dist_flag -pi $power_index \
         -sf $split_flag -sr $split_ratio -ps $pmin_split \
-        -tf $track_particle_flag -ns $nptl_selected -ni $nsteps_interval \
+        -tf $track_particle_flag -ptf $particle_tags_file \
+        -ni $nsteps_interval \
         -dd $diagnostics_directory -cf $conf \
-        -ld $local_dist -de $dump_escaped \
+        -ld $local_dist -ded $dump_escaped_dist -de $dump_escaped \
         -nm $nptl_max -in $inject_new_ptl -ij $inject_large_jz \
-        -sn $inject_same_nptl -ip $inject_part_box -jz $jz_min \
-        -nn $ncells_large_jz_norm -ib $inject_large_db2 \
+        -sn $inject_same_nptl -tti $tmax_to_inject -ip $inject_part_box \
+        -jz $jz_min -nn $ncells_large_jz_norm -ib $inject_large_db2 \
         -db2 $db2_min -nb $ncells_large_db2_norm \
         -iv $inject_large_divv -dv $divv_min \
         -nv $ncells_large_divv_norm -xs $ptl_xmin -xe $ptl_xmax \
@@ -158,10 +172,13 @@ stochastic () {
     duu0=5578.445 # not used for Parker transport
     kret=0.01  # kperp/kpara
     particle_v0=17.20195  # particle speed / velocity normalization (not used for Parker transport)
+    drift_param1=850964.408   # parameter #1 for particle drift in 3D
+    drift_param2=13575468.975 # parameter #2 for particle drift in 3D
+    charge=-1 # charge in unit charge
     diagnostics_directory=data/$2/transport_test_run/
     run_stochastic $momentum_dependency $pindex $mag_dependency \
         $dir_mhd_data $kpara0 $kret $diagnostics_directory $tau0 \
-        $duu0 $particle_v0
+        $duu0 $particle_v0 $drift_param1 $drift_param2 $charge
 }
 
 mhd_run_dir=/pscratch/sd/x/xiaocan/test/transport_test/athena_reconnection_test/bin_data/
